real time data

+++ app/src/main/java/com/example/pizzaorderapp/ui/screens/choose/components/Raiting.kt
@@ -11,7 +11,7 @@ import com.example.pizzaorderapp.R
 import com.example.pizzaorderapp.ui.theme.Yellow
 
 @Composable
-fun Raiting(r: Int) {
+fun Raiting(r: Float) {
     Box(
         Modifier
             .wrapContentHeight()
